Starfield Demo Official folder
Making a folder for the "official" version of the starfield demo. The version you see in the main folder will just be for testing purposes, to ensure compatibility and features being kept. Source control or not, I want there to be a source of truth on each individual project. The physics project will come soon enough too