Camera, better childoffset
Generalized a case of setting index offset on child objects. Not 100% where it could go, but good enough for now. Also added a new kind of camera that works better with standard camera controls. Will want to clean up a ton and make a camera class. Was there something else I was supposed to do too?